Omnibus changes to display and redisplay
There are now only two basic functions: `ebdb-display-records' and
`ebdb-redisplay-records'. The former simply fills an *EBDB* buffer
with formatted records. The latter takes a list of records to be
redisplayed, and an "action": how to redisplay them. It finds all the
records that are visible in buffers, and applies "action" to them by
calling the appropriate `ebdb-redisplay-record' method on each record.
Record marking, omitting, unloading, format toggling, etc -- all these
things are now implemented with the same system.