Rework *EBDB* buffer searching
Does not fix but does address #12

BUG: Searching on the labels of ebdb-user-field-simple instances still
doesn't work correctly, but at least it errs on the side of
permissiveness.

From a plumbing point of view, there's now a clear chain from 1)
various interactive search commands that call 2)
`ebdb-search-display', which is responsible for handling and
displaying the records found by 3) `ebdb-search' which actually does
the searching.

From a UX point of view, most of the search commands can now be called
with one of three prefix keys: "/" which searches all records and
replaces the contents of the *EBDB* buffer, "|" which filters on the
already-displayed records, and "+" which appends to the
already-displayed records.

This means that the `ebdb-append-*' stuff is gone.

Meanwhile, `ebdb-search-invert' etc still exists, but has been moved
to ebdb-com.el.