Change some of the EBDB buffer formatting defaults
* ebdb-format.el: By default, add ebdb-field-address to the :collapse
  slot of the formatters.  Also, by default format fields using the
  'normal style rather than the 'oneline style.

What this means, practically, is that most fields are just output
using `ebdb-string'.  Previously, everything was output as 'oneline,
which meant a bunch of spurious `split-strings', now they just dump
their whole strings.  Nothing actually specializes on 'normal, but it
will hit the catch-all.

Addresses are now 'collapse, which is re-routed to 'oneline by
default, which means addresses are the only fields that get
`split-string'ed, but it's also possible to catch 'collapse and do
something different.