Fill out functionality of gnorb messages field
Though the "notice" plumbing isn't actually hooked up yet! Actually
this should be disconnected from Gnorb altogether (in fact it already
is, there's no use of Gnorb in this file at all), and the field should
be made capable of collecting and following links from any MUA, not
just Gnus. We'll get there.

* ebdb-gnorb.el (gnorb-ebdb): New group.
  (gnorb-ebdb-collect-N-messages, gnorb-ebdb-define-recent,
  gnorb-ebdb-message-format): Customization options all stolen from
  Gnorb.
  (gnorb-ebdb-follow-link): New function for following a link.
  (ebdb-fmt-field, ebdb-string): Functions for formatting field
  contents.
  (ebdb-notice-field): Notice field method for storing links.