Change ebdb-record-change-name to accept strings
The fact that person and organization record types take different name
classes is still causing a little difficulty. We address this by
allowing ebdb-record-change-name to accept a string argument, meaning
that the parsing can happen within the record-class-specific method.
That should remove most of the leaky abstractions from various points
in the code.

* ebdb.el (ebdb-record-change-name): Provide a generic, for eldoc
  purposes. Provide person- and organization-specific methods that can
  accept strings. Take care of business inside those methods. Also,
  only set the 'name-string cache slot within the base method.
* ebdb-com.el (ebdb-edit-field): Adjust for new system -- it's
  necessary to read the new name here: do it with the right class, and
  the old instance as default.
* ebdb-mua.el (ebdb-annotate-message): Destroy "name-slot" with
  maximum prejudice.