Be more lenient about finding records during MUA auto updating
* ebdb-mua.el (ebdb-annotate-message): The logic of ignoring
mail-or-name seems backwards.  `mail-extract-address-components'
always returns a mail string, so the first comment is wrong.  If it
can't find a valid mail, it returns the same string as both the name
and mail.  Therefore, in all cases I can think of, we should be
ignoring the mail (not the name) if the strings are the same.