Don't offer to create contacts twice when sending messages
* ebdb-message.el (ebdb-after-load-hook): We run
`ebdb-message-auto-update' twice, in the message-header-setup-hook and
when sending the message. If the user has ebdb-message-auto-update-p
returning 'query or something equivalent, they could be prompted
multiple times to create new records. Hard-code the first call to use
'existing, so potential record creation only happens once, when the
user is ready to send the message.
(ebdb-message-display-only): Totally unnecessary extra function to do
that. I just don't like lambdas in hooks.