Rework Gnus/Message window configuration customization, bump version
The fundamental problem was that EBDB shouldn't have been taking over
window configuration for Gnus and message-mode.  The current
implementation provides sane defaults with no config, plus the option
to better integrate with Gnus' own window configuration system.

* ebdb-gnus.el (ebdb-gnus-window-configuration): Change behavior of
option.
(ebdb-insinuate-gnus): Only add the option value to
ebdb-window-to-buffer if it's a symbol.
* ebdb-message.el (ebdb-message-window-configuration,
ebdb-insinuate-mail): Do the same for message.
(ebdb-message-reply-yank-window-config): Obsolete.
* ebdb.info: Document.
* ebdb.el: Bump version to 0.6.23.
* ebdb-com.el (ebdb-pop-up-window): Refactor to more completely
short-circuit if the buffer is already displayed. Gnus will have
already popped up the window by now, so do a minimum of work.