Rework diary integration, release version 0.8
The nutshell: before, we dynamically injected entries into the diary
list using `diary-add-to-list'.  That only kicked in when listing
diary entries for a day, not when marking the calendar.  We now do it
"the right way", by providing a `ebdb-diary-anniversaries' function
that users can put in their diary file (or Org files).

* ebdb.el (ebdb-use-diary): Obsolete this option. Users now "use the
diary" by explicitly putting our function in their diary file.
(ebdb-diary-entries): Now a hash table, keyed on day-month pairs.
(ebdb-diary-anniversaries): New function for use in diary or Org
files. Checks the dynamic variable `date', and produces diary entries
for all anniversaries on that date.
(ebdb-init-field, ebdb-delete-field): Change the kind of data that's
pushed onto `ebdb-diary-entries'.
(ebdb-field-anniv-diary-entry): Simplify this function to just return
the string.
* ebdb.org: Note change in manual.